REFACTOR: move substract list to package
